Refactor student, teacher, apprentice-teacher classes to consolidate some of the methods.
