+++ apprentice_teacher.rb
@@ -1,45 +1,22 @@
-class ApprenticeTeacher
+require_relative 'school_person'
+require_relative 'module'
+
+class ApprenticeTeacher < SchoolPerson
+  include Raise
+  include Teach
   attr_reader :age, :salary, :phase, :target_raise
   attr_accessor :name
 
-  def initialize(options={})
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+  def initialize(options = {})
+    super
     @target_raise = 800
-    @phase = 3
-  end
-
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works. "
-    response += "*drops crazy knowledge bomb* "
-    response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
     response = ""
     if rating > 80
-      response = "Yay, I'm a great employee!"
       receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
     else
       response = "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
     end
