+++ teacher.rb
@@ -1,38 +1,15 @@
-class Teacher
+require_relative 'school_person'
+require_relative 'module'
+
+class Teacher < SchoolPerson
+  include Raise
+  include Teach
   attr_reader :age, :salary, :phase, :performance_rating, :target_raise
   attr_accessor :name
 
-  def initialize(options={})
-    @phase = 3
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+  def initialize(options = {})
+    super
     @target_raise = 1000
-  end
-
-  def offer_high_five
-    "High five!"
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works, fo SHO! "
-    response += "*drops flat-out insane knowledge bomb* "
-    response += "... You're welcome. *saunters away*"
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
   end
 
   def set_performance_rating(rating)
@@ -41,8 +18,8 @@ class Teacher
       receive_raise(@target_raise)
       response = "Yay, I'm a great employee!"
     else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
+      response += "Oh, well -- thanks to this actionable, specific, and kind feedback, I'll do better next time."
+      # response += ""
     end
     response
   end
